sdhci: Move common part to a separate module
