+++ src/drivers/mmc/host/sdhci/Mybuild
@@ -1,5 +1,11 @@
 package embox.driver.mmc.host
 
+static module sdhci_ops {
+	option number log_level=1
+
+	source "sdhci_ops.c"
+}
+
 module sdhci {
 	option number log_level=1
 
@@ -9,8 +15,21 @@ module sdhci {
 
 	source "sdhci.c"
 
+	depends sdhci_ops
 	depends embox.driver.block_common
 	depends embox.driver.mmc.mmc_core
+}
+
+module imx_usdhc {
+	option number log_level=1
+
+	option number base_addr
+	option number irq_num
+	option number hc_quantity
+
+	source "imx_usdhc.c"
+
+	depends sdhci_ops
 }
 
 module imx6_sdhci {
