+++ src/drivers/mmc/host/sdhci/sdhci_ops.h
@@ -1,0 +1,16 @@
+/**
+ * @file sdhc_ops.h
+ * @brief
+ * @author Denis Deryugin <deryugin.denis@gmail.com>
+ * @version
+ * @date 30.10.2019
+ */
+
+#ifndef SDHC_OPS_H
+#define SDHC_OPS_H
+
+#include "sdhci.h"
+
+extern int sdhci_hw_probe(struct sdhci_host *host);
+
+#endif /* SDHC_OPS_H */
